refactor: remove v1
Fixes #338
