+++ .github/workflows/docker_smoketest.yml
@@ -30,6 +30,6 @@ jobs:
       - name: Run integration test
         shell: bash
         run: |
-          TEST_PATH=./personal_mnemonic_medium/tests/docker_smoketest.sh
+          TEST_PATH=docker_smoketest.sh
           chmod +x $TEST_PATH
           $TEST_PATH
