+++ docker_smoketest.sh
