--- personal_mnemonic_medium/data_access/exporters/anki/card_types/test_cards.md
@@ -1,14 +1,0 @@
-Q. A card with a GUID.
-A. And here is its answer.
-
-QS. How about a card like this?
-A. Yes, an answer too.
-
-Q. How about multiline questions?
-* Like this
-* Or this?
-A. What is the hash?
-
-And some {cloze} deletions? For sure! Multipe {even}.
-
-<!-- {BearID:7696CDCD-803A-40BC-88D8-855DDBEC56CA-31546-000054DF17EAE2C1} -->
