--- personal_mnemonic_medium/data_access/exporters/anki/exporter.py
@@ -1,151 +1,0 @@
-"""This class takes a bunch of prompts, packages them, then syncs them to Anki.
-
-Can take an arbitrary amount of post-processing steps to be applied.
-"""
-
-import logging
-import traceback
-from collections.abc import Sequence
-from dataclasses import dataclass
-from pathlib import Path
-from typing import Any
-
-from wasabi import Printer
-
-from personal_mnemonic_medium.data_access.exporters.anki.sync.anki_gateway import (
-    get_anki_server_guid2noteinfo,
-)
-from personal_mnemonic_medium.data_access.exporters.anki.sync.bundle_generator import (
-    AnkiPackageGenerator,
-    DeckBundle,
-)
-from personal_mnemonic_medium.data_access.exporters.base import (
-    PromptExporter,
-)
-from personal_mnemonic_medium.domain.prompt_extractors.prompt import (
-    Prompt,
-)
-
-from .sync.gateway_utils import AnkiConnectParams, invoke
-
-log = logging.getLogger(__name__)
-# Log to disk, not to console.
-logging.basicConfig(
-    filename="anki_package_generator.log",
-    filemode="w",
-    level=logging.DEBUG,
-)
-msg = Printer(timestamp=True)
-
-
-@dataclass(frozen=True)
-class NoteDiff:
-    deck_bundle: DeckBundle
-    deckbundle_guids: set[str]
-    anki_guids: set[str]
-
-    @property
-    def symmetric_diff(self) -> set[str]:
-        return self.deckbundle_guids.symmetric_difference(
-            self.anki_guids
-        )
-
-    @property
-    def guids_only_in_anki(self) -> set[str]:
-        return self.anki_guids - self.deckbundle_guids
-
-
-# TODO: https://github.com/MartinBernstorff/personal-mnemonic-medium/issues/272 refactor: create pydantic class for noteinfo
-# TODO: https://github.com/MartinBernstorff/personal-mnemonic-medium/issues/273 Add AnkiConnect gateway, using all stateful operations
-class AnkiExporter(PromptExporter):
-    """Generates an anki package from a list of anki cards"""
-
-    def __init__(
-        self, anki_connect: AnkiConnectParams, write_apkg_dir: Path
-    ) -> None:
-        self.write_apkg_dir = write_apkg_dir
-        self.anki_connect = anki_connect
-
-    def get_anki_database_note_infos(
-        self, deck_bundle: DeckBundle
-    ) -> dict[str, Any]:
-        return get_anki_server_guid2noteinfo(deck_bundle)
-
-    def get_note_diff(self, deck_bundle: DeckBundle) -> NoteDiff:
-        anki_note_info_by_guid = self.get_anki_database_note_infos(
-            deck_bundle
-        )
-        anki_note_guids: set[str] = set(anki_note_info_by_guid.keys())
-
-        return NoteDiff(
-            deckbundle_guids=deck_bundle.note_guids,
-            anki_guids=anki_note_guids,
-            deck_bundle=deck_bundle,
-        )
-
-    def add_to_anki(
-        self, deck_bundle: DeckBundle, write_apkg_dir: Path
-    ) -> None:
-        """Adds a deck bundle to anki, and syncs it to the anki server.
-
-        Takes two paths to handle containerisation. Within the container, write to write_apkg_path. Outside of the container, import from import_apkg_from_path.
-
-        Args:
-            deck_bundle (DeckBundle): The deck bundle to add to anki
-            write_apkg_dir (Path): The path to write the apkg to
-        """
-        apkg_name = "deck.apkg"
-
-        container_path = write_apkg_dir / apkg_name
-        deck_bundle.save_to_apkg(output_path=container_path)
-
-        deck_name: str = deck_bundle.deck.name  # type: ignore
-        sync_path = self.anki_connect.apkg_dir / apkg_name
-        try:
-            invoke("importPackage", path=sync_path)
-            print(f"Imported {deck_name}!")
-        except Exception:
-            print(
-                f"""Unable to sync {deck_name}.
-    Path written to: {container_path}
-    Path to sync from: {sync_path}
-"""
-            )
-            traceback.print_exc()
-
-    def delete_diff(self, note_diff: NoteDiff):
-        try:
-            guids_to_delete = note_diff.guids_only_in_anki
-
-            if guids_to_delete:
-                guid2noteinfo = get_anki_server_guid2noteinfo(
-                    deck_bundle=note_diff.deck_bundle
-                )
-
-                note_ids = [  # type: ignore
-                    guid2noteinfo[guid]["noteId"]  # type: ignore
-                    for guid in guids_to_delete
-                ]
-
-                invoke("deleteNotes", notes=note_ids)
-                msg.good(f"Deleted {len(guids_to_delete)} notes")
-
-        except Exception:
-            msg.fail(
-                f"Unable to delete cards in {note_diff.deck_bundle.deck.name}"  # type: ignore
-            )
-            # Print full stack trace
-            traceback.print_exc()
-
-    def sync_prompts(self, prompts: Sequence[Prompt]):
-        bundles = AnkiPackageGenerator().prompts_to_bundles(prompts)
-        for bundle in bundles:
-            note_diff = self.get_note_diff(deck_bundle=bundle)
-            if len(note_diff.symmetric_diff) > 0:
-                self.add_to_anki(
-                    deck_bundle=bundle,
-                    write_apkg_dir=self.write_apkg_dir,
-                )
-
-                if self.anki_connect.delete_cards:
-                    self.delete_diff(note_diff=note_diff)
