--- personal_mnemonic_medium/data_access/exporters/base.py
@@ -1,11 +1,0 @@
-from collections.abc import Sequence
-from typing import Protocol
-
-from personal_mnemonic_medium.domain.prompt_extractors.prompt import (
-    Prompt,
-)
-
-
-class PromptExporter(Protocol):
-    def sync_prompts(self, prompts: Sequence[Prompt]):
-        ...
