--- personal_mnemonic_medium/docs_example.py
@@ -1,16 +1,0 @@
-"""
-This is just a sample script used for showcasing the documentation.
-"""
-
-
-def sample_function(arg1: str, arg2: str) -> str:
-    """This is a sample function.
-
-    Args:
-        arg1: This is the first argument.
-        arg2: This is the second argument.
-
-    Returns:
-        This is the return value.
-    """
-    return arg1 + arg2
