--- personal_mnemonic_medium/domain/prompt_extractors/test_cloze_extractor.py
@@ -1,39 +1,0 @@
-from pathlib import Path
-
-from personal_mnemonic_medium.data_access.document_ingesters.document import (
-    Document,
-)
-from personal_mnemonic_medium.domain.prompt_extractors.cloze_extractor import (
-    ClozePromptExtractor,
-)
-
-
-def test_cloze_with_hits():
-    note_with_cloze = Document(
-        content=r"""Test content.
-{Cloze deletions} are part of the {game}. Right?
-Even after newlines.
-
-Even after new blocks. They should {all} be found.
-        """,
-        uuid="1234",
-        source_path=Path(__file__),
-    )
-
-    prompts = ClozePromptExtractor().extract_prompts(note_with_cloze)
-
-    assert len(prompts) == 3
-
-
-def test_cloze_no_hits():
-    note_without_cloze = Document(
-        content=r"""Test content.""",
-        uuid="1234",
-        source_path=Path(__file__),
-    )
-
-    prompts = ClozePromptExtractor().extract_prompts(
-        note_without_cloze
-    )
-
-    assert len(prompts) == 0
