--- personal_mnemonic_medium/presentation/cli.py
@@ -1,6 +1,0 @@
-import typer
-
-from personal_mnemonic_medium.main import main
-
-if __name__ == "__main__":
-    typer.run(main)
