--- personal_mnemonic_medium/test_main.py
@@ -1,19 +1,0 @@
-from pathlib import Path
-
-from .main import main
-
-
-def test_smoke_main(tmpdir: Path):
-    apkg_output_filepath = (
-        Path(tmpdir) / "subdir" / "output_deck.apkg"
-    )
-
-    main(
-        input_dir=Path(__file__).parent / "tests",
-        apkg_output_filepath=apkg_output_filepath,
-        host_ankiconnect_dir=apkg_output_filepath,
-        watch=False,
-        use_anki_connect=False,
-    )
-
-    assert apkg_output_filepath.exists()
