--- personal_mnemonic_medium/tests/test.md
@@ -1,48 +1,0 @@
-# FSH
-
-Follikel-stimulerende hormon.
-
-QS. Hvilke biokemiske faktorer *leder til* sekretion af [[FSH]]? 
-A. [[GnRH]]
-
-QS. Hvilke biokemiske faktorer *hæmmer* sekretion af [[FSH]]? 
-A. [[Testosteron]]/[[Østrogen]] og [[Inhibin]]
-
-QG. Hvilke(t) organ(er) påvirkes af [[FSH]]? 
-A. Ovarier/tests
-
-{And now} for some testing of cloze deletions – how much does this {matter}? More than me just accepting that it's {missing} {deletions}.
-
-## Backlinks
-
-* [[Oligomenorre]]
-  * Q. Din patient viser symptomer på [[Oligomenorre]]. Hvad vil du tilføje til *paraklinikken*? 
-* [[FSH]]
-  * Q. Hvilke biokemiske faktorer *leder til* sekretion af [[FSH]]? 
-  * Q. Hvilke biokemiske faktorer *hæmmer* sekretion af [[FSH]]? 
-  * Q. Hvilke(t) organ(er) påvirker [[FSH]]? 
-  * Q. Hvilken effekt har [[FSH]]? 
-  * Q. Hvilken sygdom leder *hypo*sekretion af [[FSH]] til? 
-  * Q. Hvad er symptomerne på *hypo*sekretion af [[FSH]] i mænd? 
-    Dvs. svarende til [[Hypogonadisme]].
-  * Q. Hvilke anamnestiske faktorer vil tyde på *hypo*sekretion af [[FSH]] i kvinder? 
-  * Q. Hvordan diagnosticeres *hypo*sekretion af [[FSH]] i mænd? 
-  * Q. Hvordan diagnosticeres *hypo*sekretion af [[FSH]] i præmenopausale kvinder? 
-  * Q. Hvordan diagnosticeres *hypo*sekretion af [[FSH]] i postmenupausale kvinder? 
-* [[Infertilitet]]
-  * Q. Din patient har [[Infertilitet]] i AP. Hvad vil du tilføje til *paraklinikken* for /kvinden/? 
-* [[§Hypofysen]]
-  * [[ACTH]]
-    [[Prolaktin]]
-    [[TSH]]
-    [[LH]]
-    [[FSH]]
-    [[ADH]]
-* [[GnRH]]
-  * Fører til frigivelse af [[FSH]]/[[LH]]
-* [[Hyperprolaktinæmi]]
-  * Q. Hvorfor giver [[Hyperprolaktinæmi]] infertilitet?
-
-<!-- #anki/tag/med/Endocrinology #anki/deck/Medicine -->
-
-<!-- {BearID:91BC55B8-CB70-4ED8-A1A3-E7320FE296E7-966-000015D2D4AA76EE} -->
