--- personal_mnemonic_medium/tests/test_md_files/Bensår.md
@@ -1,84 +1,0 @@
-# Bensår
-![](BearImages/C1C47096-5A13-4BC6-BDC6-CC0F126EA691-62499-00007BA30DD2A745/23B8BF26-1E95-4DCD-AEBC-34722EB6CB27.png)
-
-	[[Bensår pga. venøs insufficiens]]
-
-QG. Hvilke overordnede kategorier [[Bensår]] findes?
-A. 1) Venøse, 2) Arterielle, 3) Immunologisk
-
-QG. Hvilke immunologiske bensår findes?
-A. 1) [[Pyoderma gangraenosum]], 2) [[Vasculitissår]]
-
-## Generelt
-
-## Differentialdiagnose
-QG. Hvad er de hyppigste årsager til [[Bensår]]?
-A. Arteriel- og venøs insufficiens.
-
-En patient har et bensår. Klinisk kan du askilles [[Bensår pga. venøs insufficiens]] fra [[Iskæmiske bensår]] ved at [[Iskæmiske bensår]] 
-* S) {Sidder perifært} 
-* C) {er skarpt afgrænsede og mindre}, 
-* A) {Smerterne bliver bedre ved at have “benet liggende ud over sengekanten”} og 
-* S) {Kun de iskæmiske gør ondt}. 
-
-QG. Din patient har et sår på benet. Hvordan vil du nærme dig, om der er tale om et  [[Vasculitissår]] eller [[Iskæmiske bensår]]?
-A. 1) Kendt morbiditet, 2) [[Vasculitissår]] starter som palpabel [[Purpura]] og 3) er næsten altid bilaterale
-
-QG. Hvilke *vaskulære* differentialdiagnoser findes til [[Bensår]]?
-A. 1) [[Hjerteinsufficiens]], 2) Arteriel- og venøs insufficiens
-
-QG. Hvilke *traumatiske* differentialdiagnoser findes til [[Bensår]]?
-A. [[Tryksår]], Traume
-
-QG. Hvilke *autoimmune* differentialdiagnoser findes til [[Bensår]]?
-A. [[Pyoderma gangraenosum]], [[Panniculitis]], [[Vasculitis]] 
-
-QG. Hvilke *endokrine* differentialdiagnoser findes til [[Bensår]]?
-A. [[§Diabetes]]
-
-
-## Udredning
-### Anamnese
-QG. Hvilke faktorer skal der spørges til specifikt for [[Bensår]]?
-A. 1) [[§Diabetes]], 2) Kendt hjertesvigt el. symptomer på [[Vaskulær claudicatio]], 3) Symptomer på [[Dyb thromboflebitis]], 4) [[Stivkrampevaccination]]
-
-### Objektiv us.
-QG. Hvilke dele i den objektive us. er specifikt for [[Bensår]]?
-A. 1) Tegn til ødem, 2) [[a. dorsalis pedis]] og [[a. tibialis post.]] palpation, 3) Temperatur og behåring af ben og fødder, 4) Infektionstegn i såret, 5) Perifær sensibilitet
-
-### Paraklinik
-
-## Behandling
-
-
-## Opfølgning
-
-
-## Prognose
-
-
-
-
-
-
-
-
-
-
-
-
-
-Sår på [[§Ben]].
-
-## Backlinks
-* [[Bensår]]
-	* QG. Hvilke overordnede kategorier [[Bensår]] findes?
-	* QG. Hvad er de hyppigste årsager til [[Bensår]]?
-	* QG. Hvilke faktorer skal der spørges til specifikt for [[Bensår]]?
-	* QG. Hvilke dele i den objektive us. er specifikt for [[Bensår]]?
-* [[Pyoderma gangraenosum]]
-	* [[Bensår]]
-
-<!-- #anki/tag/med/Derma #anki/deck/Medicine #anki/tag/med/Endocrinology -->
-
-<!-- {BearID:4a6a4cf2e514bc0c0844cd4b6b03a68c} -->
