--- personal_mnemonic_medium/tests/test_md_files/test_header_removal.md
@@ -1,15 +1,0 @@
-# This is a header
-
-Q. And this is a prompt
-A. Yes
-
-# Now a new header
-Q. And another prompt
-A. Has the header been removed from the card?
-<!-- BearID:03927591-B032-4599-90FC-E1F6179D6529-14366-0000028A1B470030 -->
-<!-- BearID:03927591-B032-4599-90FC-E1F6179D6529-14366-0000028A1B470030 -->
-<!-- BearID:03927591-B032-4599-90FC-E1F6179D6529-14366-0000028A1B470030 -->
-<!-- BearID:03927591-B032-4599-90FC-E1F6179D6529-14366-0000028A1B470030 -->
-<!-- BearID:03927591-B032-4599-90FC-E1F6179D6529-14366-0000028A1B470030 -->
-<!-- BearID:03927591-B032-4599-90FC-E1F6179D6529-14366-0000028A1B470030 -->
-<!-- {BearID:f551f34a2e6c3ea0467a1b8fe897d6b8e0fdc413747c3c465745371aba2b9b2c} -->
