--- personal_mnemonic_medium/tests/test_md_files/test_newlines_in_lists.md
@@ -1,18 +1,0 @@
-# Test_file
-
-Q. This is a question with newlines:
-* Point 1
-* Point 2
-A. And an answer
-
-Q. Question without newlines
-A. But answers with them!
-* Answer 1
-* Answer 2
-<!-- BearID:03927591-B032-4599-90FC-E1F6179D6529-14366-0000028A1B470030 -->
-<!-- BearID:03927591-B032-4599-90FC-E1F6179D6529-14366-0000028A1B470030 -->
-<!-- BearID:03927591-B032-4599-90FC-E1F6179D6529-14366-0000028A1B470030 -->
-<!-- BearID:03927591-B032-4599-90FC-E1F6179D6529-14366-0000028A1B470030 -->
-<!-- BearID:03927591-B032-4599-90FC-E1F6179D6529-14366-0000028A1B470030 -->
-<!-- BearID:03927591-B032-4599-90FC-E1F6179D6529-14366-0000028A1B470030 -->
-<!-- {BearID:57a596cc5512b2466c97dda139237ebe755f592453824bb38c2664f8e27ac604} -->
