+++ personal_mnemonic_medium/v2/domain/prompt_destination/anki_connect/anki_prompt_css.py
