+++ personal_mnemonic_medium/v2/presentation/cli.py
@@ -1,16 +1,22 @@
 import logging
+import os
 from pathlib import Path
 from typing import Annotated
 
 import sentry_sdk
 import typer
 
-from personal_mnemonic_medium.main import get_env
 from personal_mnemonic_medium.v2.sync_deck import sync_deck
 
 log = logging.getLogger(__name__)
 
 app = typer.Typer()
+
+
+def get_env(default: str) -> str:
+    if os.getenv("GITHUB_ACTIONS"):
+        return "GITHUB_ACTIONS"
+    return os.getenv("ENV", default)
 
 
 @app.command()
